fix sidebar hook and modularize projects page
