+++ src/components/layout/Sidebar.tsx
@@ -4,7 +4,7 @@ import { Link, useLocation, useNavigate } from 'react-router-dom';
 import type { ReactNode } from 'react';
 
 import { useUIStore } from '@/stores/uiStore';
-import { useCategories } from '@/hooks/useProjects';
+import { useCategories } from '@/hooks/useCategories';
 import { navItems } from './Sidebar/constants';
 
 import { CategoryItem } from './Sidebar/CategoryItem';
